+++ examples/04-frontier-faces/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cotal-ai/example-04-frontier-faces
 
+## 0.0.77
+
+### Patch Changes
+
+- @cotal-ai/core@0.47.1
+- @cotal-ai/workspace@0.47.1
+
+## 0.0.76
+
+### Patch Changes
+
+- @cotal-ai/core@0.47.0
+- @cotal-ai/workspace@0.47.0
+
 ## 0.0.75
 
 ### Patch Changes