+++ examples/05-scale-showcase/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cotal-ai/example-05-scale-showcase
 
+## 0.0.35
+
+### Patch Changes
+
+- @cotal-ai/core@0.47.1
+
+## 0.0.34
+
+### Patch Changes
+
+- @cotal-ai/core@0.47.0
+
 ## 0.0.33
 
 ### Patch Changes