+++ examples/06-feed-agent/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cotal-ai/example-06-feed-agent
 
+## 0.0.16
+
+### Patch Changes
+
+- @cotal-ai/core@0.47.1
+
+## 0.0.15
+
+### Patch Changes
+
+- @cotal-ai/core@0.47.0
+
 ## 0.0.14
 
 ### Patch Changes